[EGE #108] String class refactored. Unittests added.

+++ Sources/Core/Tools/CommandLineParser.h
@@ -29,14 +29,14 @@ class CommandLineParser
      *  @param  pos         Position within commandline from which the search should be done.
      *  @return Position within commandline where the token has been found. Returns String::npos if nothing was found.
      */
-    size_t findNextSwitchToken(const String& commandLine, size_t pos);
+    s32 findNextSwitchToken(const String& commandLine, s32 pos);
     /*! Retrieves the name from the command line.
      *  @param  commandLine String with entire command line.
      *  @param  pos         Position within commandline from which name should be retrieved.
      *  @return Name found at given location.
      *  @note Name is the first n-chracters found before SPACE, EQUAL appears or end of commanline is reached.
      */
-    String retrieveName(const String& commandLine, size_t pos);
+    String retrieveName(const String& commandLine, s32 pos);
 
   private:
